fixup: Move into new _NIOBeta57 module
