+++ Package.swift
@@ -36,6 +36,8 @@ var targets: [PackageDescription.Target] = [
                            "NIOPosix"]),
     .target(name: "_NIOConcurrency",
             dependencies: ["NIO", "NIOCore"]),
+    .target(name: "_NIOBeta57",
+            dependencies: ["NIOCore"]),
     .target(name: "NIOFoundationCompat", dependencies: ["NIO", "NIOCore"]),
     .target(name: "CNIOAtomics", dependencies: []),
     .target(name: "CNIOSHA1", dependencies: []),
@@ -91,7 +93,7 @@ var targets: [PackageDescription.Target] = [
     .executableTarget(name: "NIOAsyncAwaitDemo",
             dependencies: ["NIOPosix", "NIOCore", "NIOHTTP1"]),
     .testTarget(name: "NIOCoreTests",
-                dependencies: ["NIOCore", "NIOEmbedded", "NIOFoundationCompat"]),
+                dependencies: ["NIOCore", "NIOEmbedded", "NIOFoundationCompat", "_NIOBeta57"]),
     .testTarget(name: "NIOEmbeddedTests",
                 dependencies: ["NIOConcurrencyHelpers", "NIOCore", "NIOEmbedded"]),
     .testTarget(name: "NIOPosixTests",
@@ -122,6 +124,7 @@ let package = Package(
         .library(name: "NIOEmbedded", targets: ["NIOEmbedded"]),
         .library(name: "NIOPosix", targets: ["NIOPosix"]),
         .library(name: "_NIOConcurrency", targets: ["_NIOConcurrency"]),
+        .library(name: "_NIOBeta57", targets: ["_NIOBeta57"]),
         .library(name: "NIOTLS", targets: ["NIOTLS"]),
         .library(name: "NIOHTTP1", targets: ["NIOHTTP1"]),
         .library(name: "NIOConcurrencyHelpers", targets: ["NIOConcurrencyHelpers"]),
